[gpu-video] Support multiple backends on `VideoInstance` and `VideoAdapter`
