+++ Cargo.toml
@@ -37,7 +37,7 @@ smelter-api = { path = "smelter-api" }
 smelter-core = { path = "smelter-core" }
 smelter-render = { path = "smelter-render" }
 libcef = { path = "libcef" }
-gpu-video = { path = "gpu-video", features = ["expose-parsers"] }
+gpu-video = { path = "gpu-video", features = ["expose-parsers", "expose-backends"] }
 rtmp = { path = "rtmp" }
 bytes = "1.11.1"
 serde_json = { version = "1.0.99", features = ["preserve_order"] }
