+++ gpu-video/src/backends.rs
@@ -1,0 +1,2 @@
+#[cfg(vulkan)]
+pub mod vulkan;
