+++ gpu-video/src/backends/vulkan.rs
@@ -1,0 +1,5 @@
+pub(crate) mod vulkan_adapter;
+pub(crate) mod vulkan_instance;
+
+pub use vulkan_adapter::{VulkanAdapter, VulkanAdapterInfo, VulkanAdapterInitError};
+pub use vulkan_instance::{VulkanInstance, VulkanInstanceInitError};
