+++ gpu-video/src/instance.rs
@@ -1,11 +1,9 @@
-use std::{ffi::CStr, sync::Arc};
-
-use ash::{Entry, vk};
+use std::sync::Arc;
 
 use crate::{
-    VideoInitError,
+    VideoInstanceInitError,
     adapter::{VideoAdapter, VideoAdapterDescriptor},
-    wrappers::*,
+    backends,
 };
 
 /// Describes a [`VideoInstance`].
@@ -18,135 +16,51 @@ pub struct VideoInstanceDescriptor {
     pub enable_api_dump: bool,
 }
 
+pub(crate) trait VideoInstanceBackend {
+    fn iter_adapters<'a>(
+        &'a self,
+    ) -> Result<Box<dyn Iterator<Item = VideoAdapter<'a>> + 'a>, VideoInstanceInitError>;
+}
+
 /// Context for all encoders and decoders.
+#[derive(Clone)]
 pub struct VideoInstance {
-    _entry: Entry,
-    pub(crate) instance: Arc<Instance>,
-    _debug_messenger: Option<DebugMessenger>,
+    instance: Arc<dyn VideoInstanceBackend>,
 }
 
 impl VideoInstance {
-    pub fn new(desc: &VideoInstanceDescriptor) -> Result<Arc<Self>, VideoInitError> {
-        let entry = unsafe { Entry::load()? };
-        Self::new_from_entry(entry, &mut Vec::new(), desc)
-    }
-
-    #[cfg(vulkan)]
-    pub fn new_from_entry(
-        entry: Entry,
-        extensions: &mut Vec<&'static CStr>,
-        desc: &VideoInstanceDescriptor,
-    ) -> Result<Arc<Self>, VideoInitError> {
-        let api_version = vk::make_api_version(0, 1, 3, 0);
-        let app_info = vk::ApplicationInfo {
-            api_version,
-            ..Default::default()
-        };
-
-        let mut requested_layers = Vec::new();
-        if desc.enable_validations {
-            requested_layers.push(c"VK_LAYER_KHRONOS_validation");
-        }
-        if desc.enable_api_dump {
-            requested_layers.push(c"VK_LAYER_LUNARG_api_dump");
-        }
-
-        let instance_layer_properties = unsafe { entry.enumerate_instance_layer_properties()? };
-        let instance_layer_names = instance_layer_properties
-            .iter()
-            .map(|layer| layer.layer_name_as_c_str())
-            .collect::<Result<Vec<_>, _>>()?;
-
-        let layers = requested_layers
-            .into_iter()
-            .filter(|requested_layer_name| {
-                instance_layer_names
-                    .iter()
-                    .any(|instance_layer_name| instance_layer_name == requested_layer_name)
-            })
-            .map(|layer| layer.as_ptr())
-            .collect::<Vec<_>>();
-
-        if !extensions.contains(&vk::EXT_DEBUG_UTILS_NAME) {
-            extensions.push(vk::EXT_DEBUG_UTILS_NAME);
-        }
-
-        let extension_ptrs = extensions.iter().map(|e| e.as_ptr()).collect::<Vec<_>>();
-
-        let create_info = vk::InstanceCreateInfo::default()
-            .application_info(&app_info)
-            .enabled_layer_names(&layers)
-            .enabled_extension_names(&extension_ptrs);
-
-        let instance = unsafe { entry.create_instance(&create_info, None) }?;
-        let instance = Arc::new(Instance::new(
-            instance.clone(),
-            entry.clone(),
-            desc.enable_validations,
-            true,
-        ));
-
-        let debug_messenger = if desc.enable_validations {
-            Some(DebugMessenger::new(instance.clone())?)
-        } else {
-            None
-        };
+    pub fn new(desc: &VideoInstanceDescriptor) -> Result<Self, VideoInstanceInitError> {
+        #[cfg(vulkan)]
+        let instance = backends::vulkan::VulkanInstance::new(desc)?;
 
         Ok(Self {
-            _entry: entry,
-            instance,
-            _debug_messenger: debug_messenger,
-        }
-        .into())
-    }
-
-    /// Creates an instance that does not own `ash::Instance`. The instance is not destroyed on drop.
-    #[cfg_attr(not(feature = "wgpu"), allow(dead_code))]
-    pub(crate) fn new_unowned(
-        instance: ash::Instance,
-        entry: Entry,
-        desc: &VideoInstanceDescriptor,
-    ) -> Self {
-        let instance = Arc::new(Instance::new(
-            instance.clone(),
-            entry.clone(),
-            desc.enable_validations,
-            false,
-        ));
-        Self {
-            _entry: entry,
-            instance,
-            _debug_messenger: None,
-        }
+            instance: Arc::new(instance),
+        })
     }
 
     /// Creates an adapter that meets requirements specified in the descriptor.
     pub fn create_adapter<'a>(
         &'a self,
         descriptor: &VideoAdapterDescriptor,
-    ) -> Result<VideoAdapter<'a>, VideoInitError> {
+    ) -> Result<VideoAdapter<'a>, VideoInstanceInitError> {
         self.iter_adapters()?
             .find(|adapter| {
                 (!descriptor.supports_decoding || adapter.supports_decoding())
                     && (!descriptor.supports_encoding || adapter.supports_encoding())
             })
-            .ok_or(VideoInitError::NoDevice)
+            .ok_or(VideoInstanceInitError::NoAdapterFound)
     }
 
     /// Iterator over all available [`VideoAdapter`]s that support at least decoding or encoding.
     pub fn iter_adapters<'a>(
         &'a self,
-    ) -> Result<impl Iterator<Item = VideoAdapter<'a>> + 'a, VideoInitError> {
-        crate::adapter::iter_adapters(self)
-    }
-
-    pub fn raw_instance(&self) -> ash::Instance {
-        self.instance.instance.clone()
+    ) -> Result<impl Iterator<Item = VideoAdapter<'a>>, VideoInstanceInitError> {
+        self.instance.iter_adapters()
     }
 }
 
 impl std::fmt::Debug for VideoInstance {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("VulkanInstance").finish()
+        f.debug_struct("VideoInstance").finish()
     }
 }
