+++ gpu-video/src/lib.rs
@@ -1,7 +1,16 @@
 #![doc = include_str!("../README.md")]
 
+#[cfg(feature = "expose-backends")]
+pub mod backends;
+#[cfg(all(vulkan, not(feature = "expose-backends")))]
+pub(crate) mod backends;
+
+// TODO: After caps refactor cfg for instance and adapter won't be needed
 #[cfg(vulkan)]
 mod adapter;
+#[cfg(vulkan)]
+mod instance;
+
 #[cfg(vulkan)]
 pub(crate) mod codec;
 #[cfg(vulkan)]
@@ -9,8 +18,6 @@ mod device;
 // TODO: cfg(vulkan) will not be needed once we add proper metal support
 #[cfg(all(vulkan, feature = "wgpu"))]
 mod global_registry;
-#[cfg(vulkan)]
-mod instance;
 #[cfg(vulkan)]
 mod vulkan_decoder;
 #[cfg(vulkan)]
@@ -22,6 +29,7 @@ pub(crate) mod wgpu_helpers;
 #[cfg(vulkan)]
 pub(crate) mod wrappers;
 
+// TODO: This could be inlined with lib.rs when metal support is added
 #[cfg(vulkan)]
 mod vulkan_video;
 #[cfg(vulkan)]
